voting componemt created
